refactor(web): fixes unit-test helper func name, puts in own script
